Create a substater with an `extraPrefix`
Sometimes you create a RemoteClient with a prefix of "project_name".
You then may want to pass this stater into other objects adding a prefix
identifing the that object is being used. This PR allows for a such a
"substater" to be created using the same underlying connection as the original.